agents/requirements.txt
controlflow>=0.12.1
asyncpg>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
loguru>=0.7.2
pydantic>=2.5.0
sqlalchemy>=2.0.0
asyncio-mqtt>=0.16.0